YUK-9450, Yukon should support the ability to switch the UtilityPRO in and out of AUTO mode when using the thermostat manual screen
Fixed compile error in test code.

+++ yukon-client/api/test/com/cannontech/yukon/api/loadManagement/mocks/MockAccountEventLogService.java
@@ -310,8 +310,10 @@ public class MockAccountEventLogService implements AccountEventLogService{
     }
 
     @Override
-    public void thermostatManualSetAttemptedByApi(LiteYukonUser yukonUser, String serialNumber, int heatTemperature,
-                                                  int coolTemperature, String mode, String fanSetting, boolean holdTemperature) {
+    public void thermostatManualSetAttemptedByApi(LiteYukonUser yukonUser, String serialNumber,
+                                                  Double heatTemperature, Double coolTemperature,
+                                                  String mode, String fanSetting,
+                                                  boolean holdTemperature) {
     }
 
     @Override
